Fix directory structure
