+++ app/__init__.py
@@ -1,0 +1,4 @@
+from app.face_alignment import FaceAligner
+shape_predictor_path = "data/shape_predictor_68_face_landmarks.dat"
+face_alignment = FaceAligner(shape_predictor_path)
+
