+++ config.py
@@ -1,0 +1,4 @@
+#!/usr/bin/python3
+
+UPLOAD_FOLDER = 'images/uploads'
+
